requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
httpx==0.27.2
websockets==12.0
aiofiles==24.1.0

SpeechRecognition==3.10.4
openai-whisper==20231117
soundfile==0.12.1
pydub==0.25.1

transformers==4.44.2
accelerate==0.34.2

pyttsx3==2.98
numpy==1.26.4